Button, GameVar, ScreenFormat enums instead of ints + same other type changes
Former-commit-id: b1f30b786587e1ddc323c6abb17f145715ad4a91
Former-commit-id: d4e0453f0b972db42c6abe22e1c5a5bc9fe13acf

+++ vizia_doom_src/src/vizia_screen.h
@@ -14,9 +14,9 @@
 #define VIZIA_SCREEN_BGRA32 8
 #define VIZIA_SCREEN_ABGR32 9
 
-extern int viziaScreenWidth;
-extern int viziaScreenHeight;
-extern int viziaScreenPitch;
+extern unsigned int viziaScreenWidth;
+extern unsigned int viziaScreenHeight;
+extern size_t viziaScreenPitch;
 extern size_t viziaScreenSize;
 
 void Vizia_ScreenInit();
